fixed delta time + fixed name conflict in opengl_buffer

+++ src/core/device/device.h
@@ -28,8 +28,8 @@ typedef struct Analytics
     double currentFps;
     double currentMs;
 
-    float delta;
-    float lastFrame;
+    double delta;
+    double lastFrame;
 
     // ui32 currentDrawCalls;
 } Analytics;
